==> CODE/main_analysis/00_processing_raw_acc_data.ipynb <==
import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import datetime as dt

# set working directory to the repository 
os.chdir( "C:/Users/cloftus/Documents/baboon_sleep/" )

os.chdir( "DATA/main_data" )

dat1 = pd.read_csv("Collective movement in wild baboons.csv", header=0)

dat1 = dat1[ [ 'tag-local-identifier', 'timestamp', 'eobs:acceleration-sampling-frequency-per-axis',
 'eobs:accelerations-raw'] ]

dat1 = dat1.rename( columns = {'tag-local-identifier': 'tag', 'eobs:acceleration-sampling-frequency-per-axis': 'samp_freq', 'eobs:accelerations-raw': 'acc'} )

dat1 = dat1.dropna( subset = [ 'acc' ] )

dat1[ 'timestamp' ] = pd.to_datetime( dat1[ 'timestamp' ] )

dat1 = dat1.sort_values( by = ['tag', 'timestamp'] )

seconds = list( map( lambda x: x.time().second, dat1[ 'timestamp' ] ) )

inds_to_keep = np.where( np.array( seconds ) < 12 ) 

dat = dat1.iloc[ inds_to_keep ]

hz_12 = np.linspace( 0, 1, 13 )

final_col = []

for i in range( dat.shape[0]):
    
    if len( dat.iloc[i, 3].split( ' ' ) ) == 78:
        
        final_col.append( dat.iloc[i, 3] )
    
    else:
        
        if  len( dat.iloc[i, 3].split( ' ' ) ) != 36:
            
            print( dat.iloc[i, ] , flush = True)
            final_col.append( dat.iloc[i, 3] )

        else:
            xyz = dat.iloc[i, 3].split( ' ' )
            
            x = xyz[0::3]
            y = xyz[1::3]
            z = xyz[2::3]
            
            new_x = [ int(x[0]) ]
            new_y = [ int(y[0]) ]
            new_z = [ int(z[0]) ]
                
            new_x.append( np.average( [ int( x[1]) , int( x[2] ) ], weights = [ (hz_12[2] - 0.1 ) , ( 0.1 - hz_12[1] ) ] ) )
            new_x.append( np.average( [ int( x[2]) , int( x[3] ) ], weights = [ (hz_12[3] - 0.2 ) , ( 0.2 - hz_12[2] ) ] ) )
            new_x.append( np.average( [ int( x[3]) , int( x[4] ) ], weights = [ (hz_12[4] - 0.3 ) , ( 0.3 - hz_12[3] ) ] ) )
            new_x.append( np.average( [ int( x[4]) , int( x[5] ) ], weights = [ (hz_12[5] - 0.4 ) , ( 0.4 - hz_12[4] ) ] ) )
            new_x.append( int( x[6] ) )
            new_x.append( np.average( [ int( x[7]) , int( x[8] ) ], weights = [ (hz_12[8] - 0.6 ) , ( 0.6 - hz_12[7] ) ] ) )
            new_x.append( np.average( [ int( x[8]) , int( x[9] ) ], weights = [ (hz_12[9] - 0.7 ) , ( 0.7 - hz_12[8] ) ] ) )
            new_x.append( np.average( [ int( x[9]) , int( x[10] ) ], weights = [ (hz_12[10] - 0.8 ) , ( 0.8 - hz_12[9] ) ] ) )
            new_x.append( np.average( [ int( x[10]) , int( x[11] ) ], weights = [ (hz_12[11] - 0.9 ) , ( 0.9 - hz_12[10] ) ] ) )


            new_y.append( np.average( [ int( y[1]) , int( y[2] ) ], weights = [ (hz_12[2] - 0.1 ) , ( 0.1 - hz_12[1] ) ] ) )
            new_y.append( np.average( [ int( y[2]) , int( y[3] ) ], weights = [ (hz_12[3] - 0.2 ) , ( 0.2 - hz_12[2] ) ] ) )
            new_y.append( np.average( [ int( y[3]) , int( y[4] ) ], weights = [ (hz_12[4] - 0.3 ) , ( 0.3 - hz_12[3] ) ] ) )
            new_y.append( np.average( [ int( y[4]) , int( y[5] ) ], weights = [ (hz_12[5] - 0.4 ) , ( 0.4 - hz_12[4] ) ] ) )
            new_y.append( int( y[6] ) )
            new_y.append( np.average( [ int( y[7]) , int( y[8] ) ], weights = [ (hz_12[8] - 0.6 ) , ( 0.6 - hz_12[7] ) ] ) )
            new_y.append( np.average( [ int( y[8]) , int( y[9] ) ], weights = [ (hz_12[9] - 0.7 ) , ( 0.7 - hz_12[8] ) ] ) )
            new_y.append( np.average( [ int( y[9]) , int( y[10] ) ], weights = [ (hz_12[10] - 0.8 ) , ( 0.8 - hz_12[9] ) ] ) )
            new_y.append( np.average( [ int( y[10]) , int( y[11] ) ], weights = [ (hz_12[11] - 0.9 ) , ( 0.9 - hz_12[10] ) ] ) )


            new_z.append( np.average( [ int( z[1]) , int( z[2] ) ], weights = [ (hz_12[2] - 0.1 ) , ( 0.1 - hz_12[1] ) ] ) )
            new_z.append( np.average( [ int( z[2]) , int( z[3] ) ], weights = [ (hz_12[3] - 0.2 ) , ( 0.2 - hz_12[2] ) ] ) )
            new_z.append( np.average( [ int( z[3]) , int( z[4] ) ], weights = [ (hz_12[4] - 0.3 ) , ( 0.3 - hz_12[3] ) ] ) )
            new_z.append( np.average( [ int( z[4]) , int( z[5] ) ], weights = [ (hz_12[5] - 0.4 ) , ( 0.4 - hz_12[4] ) ] ) )
            new_z.append( int( z[6]) )
            new_z.append( np.average( [ int( z[7]) , int( z[8] ) ], weights = [ (hz_12[8] - 0.6 ) , ( 0.6 - hz_12[7] ) ] ) )
            new_z.append( np.average( [ int( z[8]) , int( z[9] ) ], weights = [ (hz_12[9] - 0.7 ) , ( 0.7 - hz_12[8] ) ] ) )
            new_z.append( np.average( [ int( z[9]) , int( z[10] ) ], weights = [ (hz_12[10] - 0.8 ) , ( 0.8 - hz_12[9] ) ] ) )
            new_z.append( np.average( [ int( z[10]) , int( z[11] ) ], weights = [ (hz_12[11] - 0.9 ) , ( 0.9 - hz_12[10] ) ] ) )
            
            full_acc = [ [ m, n, p ]  for m, n, p in zip( new_x , new_y, new_z )]
            
            new_xyz = []
            for sublist in full_acc:
                for item in sublist:
                    new_xyz.append(item)
            
            new_xyz = list( map( str, new_xyz ))
            final_xyz = ' '.join(new_xyz)
            
            final_col.append( final_xyz )
            
    if (i % 100000) == 0:
        
        print( i / dat.shape[ 0 ], flush = True )

            
    

dat[ 'new_acc' ] = final_col

dat[ 'min' ] = list( map( lambda x: x.strftime("%Y-%m-%d %H:%M"), dat['timestamp'] ))

dat[ 'sec_diff' ] =  list( map( lambda x: x.total_seconds(), dat['timestamp'].diff() ) ) 

from sortedcontainers import SortedSet

tag_names = set( dat[ 'tag' ] )

final_dat = pd.DataFrame()


for tag in tag_names:
    
    print( tag )

    final_acc_col = []
    
    id_dat = dat[ dat[ 'tag' ] == tag ]
    
    min_list = SortedSet( id_dat[ 'min' ] )
    
    for minute in min_list:
    
        min_dat = id_dat[ id_dat[ 'min' ] == minute ]

        switch = 0
        
        if min_dat.shape[0] < 3:
            
            final_acc_col.append( min_dat[ 'new_acc' ].iloc[0] )
            
        else:
            
            for j in range( min_dat.shape[0] - 2 ):
                
                if( all( min_dat['sec_diff'].iloc[ [ j + 1, j + 2 ] ] == 1 ) ):
                        
                    three_secs = ' '.join( min_dat['new_acc'].iloc[ [ j, j + 1, j + 2 ] ] )
                    
                    final_acc_col.append( ' '.join( three_secs.split( ' ' )[ : 78 ] ) )
                    
                    switch = 1
                    
                    break
                    
                    
            if switch == 0:
                
                final_acc_col.append( min_dat[ 'new_acc' ].iloc[0] )

                
    temp_df = pd.DataFrame(
            {
            'tag': tag ,
            'timestamp':  min_list ,
            'acc': final_acc_col
        }
    )
    
    final_dat = pd.concat( [ final_dat, temp_df ] )



inds_to_keep = np.where( np.array( list( map( lambda x: len( x.split( ' ' ) ), final_dat['acc'] ) ) ) == 78 )

final_dat_1 = final_dat.iloc[ inds_to_keep[0] ]

final_dat_1.to_csv("all_burst_acc.csv", index = False)
